+++ magefile.go
@@ -5,6 +5,7 @@ package main
 
 import (
 	"fmt"
+	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -22,8 +23,117 @@ const (
 
 // Build basics binary to /bin directory
 func Build() error {
-	fmt.Println("Building basics binary...")
-	cmd := exec.Command("go", "build", "-o", "./bin/basics.exe", "./cmd/basics")
+	// Build version
+	versionBytes, err := os.ReadFile("VERSION")
+	if err != nil {
+		return err
+	}
+	version := strings.TrimSpace(string(versionBytes))
+
+	parts := strings.Split(version, ".")
+
+	if len(parts) != 4 {
+		// handle error: unexpected version format
+		panic("invalid version string")
+	}
+
+	fmt.Println("Building BASICS resource files for version", version)
+
+	tpl := `{
+		"RT_GROUP_ICON": {
+		  "APP": {
+			"0000": [
+			  "icon.png",
+			  "icon16.png",
+			  "icon32.png",
+			  "icon48.png",
+			  "icon64.png",
+			  "icon128.png"
+			]
+		  }
+		},
+		"RT_MANIFEST": {
+		  "#1": {
+			"0409": {
+			  "identity": {
+				"name": "basics",
+				"version": "%[1]s"
+			  },
+			  "description": "",
+			  "minimum-os": "win7",
+			  "execution-level": "as invoker",
+			  "ui-access": false,
+			  "auto-elevate": false,
+			  "dpi-awareness": "system",
+			  "disable-theming": false,
+			  "disable-window-filtering": false,
+			  "high-resolution-scrolling-aware": false,
+			  "ultra-high-resolution-scrolling-aware": false,
+			  "long-path-aware": false,
+			  "printer-driver-isolation": false,
+			  "gdi-scaling": false,
+			  "segment-heap": false,
+			  "use-common-controls-v6": false
+			}
+		  }
+		},
+		"RT_VERSION": {
+		  "#1": {
+			"0000": {
+			  "fixed": {
+				"file_version": "%[1]s",
+				"product_version": "%[1]s"
+			  },
+			  "info": {
+				"0409": {
+				  "Comments": "",
+				  "CompanyName": "ultra-sonic-28",
+				  "FileDescription": "BASIC Interpreter for old computers",
+				  "FileVersion": "%[1]s",
+				  "InternalName": "basics",
+				  "LegalCopyright": "© 2025-2026 - ultra-sonic-28 - MIT License",
+				  "LegalTrademarks": "",
+				  "OriginalFilename": "basics.exe",
+				  "PrivateBuild": "",
+				  "ProductName": "BASICS",
+				  "ProductVersion": "%[1]s",
+				  "SpecialBuild": ""
+				}
+			  }
+			}
+		  }
+		}
+	  }
+`
+	// Générer winres.json
+	os.WriteFile(
+		"./winres/winres.json",
+		[]byte(fmt.Sprintf(tpl, version)),
+		0644,
+	)
+
+	// Generate windows resource files for embbeding
+	cmd := exec.Command("go-winres", "make")
+	cmd.Run()
+
+	if err := moveFile("./", "./cmd/basics", "rsrc_windows_386.syso"); err != nil {
+		log.Fatal(err)
+	}
+	if err := moveFile("./", "./cmd/basics", "rsrc_windows_amd64.syso"); err != nil {
+		log.Fatal(err)
+	}
+
+	// Build binaire
+	fmt.Println("Building BASICS binary...")
+	builddate := "2025-02-06"
+	flags := fmt.Sprintf("-X main.Version=%s -X main.BuildDate=%s", version, builddate)
+	cmd = exec.Command(
+		"go",
+		"build",
+		"-ldflags", flags,
+		"-o", "./bin/basics.exe",
+		"./cmd/basics",
+	)
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 	return cmd.Run()
@@ -61,4 +171,26 @@ func Clean() error {
 
 		return nil
 	})
+}
+
+// Installing tools : winres
+func Tools() error {
+	cmd := exec.Command("go", "install", "github.com/tc-hib/go-winres@v0.3.1")
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+
+	fmt.Println("Installing tools...")
+	return cmd.Run()
+}
+
+func moveFile(srcDir, dstDir, name string) error {
+	srcPath := filepath.Join(srcDir, name)
+	dstPath := filepath.Join(dstDir, name)
+
+	// Ensure destination directory exists.
+	if err := os.MkdirAll(dstDir, 0o755); err != nil {
+		return err
+	}
+
+	return os.Rename(srcPath, dstPath)
 }